+++ WalletWasabi.Gui/Models/LoadWalletEntry.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Text;
-using WalletWasabi.Hwi2.Models;
+using WalletWasabi.Hwi.Models;
 
 namespace WalletWasabi.Gui.Models
 {
