+++ WalletWasabi.Tests/IntegrationTests/HwiKatas.cs
@@ -5,9 +5,9 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using WalletWasabi.Hwi2;
-using WalletWasabi.Hwi2.Exceptions;
-using WalletWasabi.Hwi2.Models;
+using WalletWasabi.Hwi;
+using WalletWasabi.Hwi.Exceptions;
+using WalletWasabi.Hwi.Models;
 using WalletWasabi.KeyManagement;
 using Xunit;
