+++ WalletWasabi/Hwi/Binaries/README.md
